updates file organization of project
