+++ src/components/calculator-form/index.jsx
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import Slider from 'material-ui/Slider';
-
 import { Col } from 'react-materialize';
-import { formatMoney } from '../utils/formatMoney';
-import '../css/index.css';
+
+import { formatMoney } from '../../utils/formatMoney';
+import '../../css/index.css';
 
 class CalculatorForm extends Component {
   render() {
